feat(android-needs-review): Rename "AutomatedChecksView*" to "ResultsView*"
The name "automated checks" is appropriate to the subset of automated test results displayed in the main view. However, it does not accurately describe items which need reivew or other manual tests. Therefore, we are renaming automated checks view to "results" view. This is intended to indicate the view for any possible test results.
