+++ src/electron/views/results/components/command-bar.scss
