+++ src/electron/views/results/components/command-bar.tsx
