+++ src/electron/views/results/components/header-section.scss
