+++ src/electron/views/results/components/header-section.tsx
