+++ src/electron/views/results/components/maximize-restore-button.tsx
