+++ src/electron/views/results/components/reflow-command-bar.tsx
