+++ src/electron/views/results/components/title-bar.scss
