+++ src/electron/views/results/components/title-bar.tsx
