+++ src/tests/unit/tests/electron/views/results/components/title-bar.test.tsx
@@ -7,8 +7,8 @@ import { BaseClientStoresHub } from 'common/stores/base-client-stores-hub';
 import { ClientStoresHub } from 'common/stores/client-stores-hub';
 import { WindowFrameActionCreator } from 'electron/flux/action-creator/window-frame-action-creator';
 import { WindowStateStoreData } from 'electron/flux/types/window-state-store-data';
-import { MaximizeRestoreButtonProps } from 'electron/views/automated-checks/components/maximize-restore-button';
-import { TitleBar, TitleBarProps } from 'electron/views/automated-checks/components/title-bar';
+import { MaximizeRestoreButtonProps } from 'electron/views/results/components/maximize-restore-button';
+import { TitleBar, TitleBarProps } from 'electron/views/results/components/title-bar';
 import { PlatformInfo } from 'electron/window-management/platform-info';
 import { shallow } from 'enzyme';
 import { Button } from 'office-ui-fabric-react';
