+++ tsconfig.strictNullChecks.json
@@ -232,9 +232,9 @@
         "./src/electron/types/left-nav-item-key.ts",
         "./src/electron/types/left-nav-item.ts",
         "./src/electron/types/test-key.ts",
-        "./src/electron/views/automated-checks/components/header-section.tsx",
-        "./src/electron/views/automated-checks/components/maximize-restore-button.tsx",
-        "./src/electron/views/automated-checks/components/title-bar.tsx",
+        "./src/electron/views/results/components/header-section.tsx",
+        "./src/electron/views/results/components/maximize-restore-button.tsx",
+        "./src/electron/views/results/components/title-bar.tsx",
         "./src/electron/views/bundled-renderer-styles.ts",
         "./src/electron/views/common/window-title/window-title.tsx",
         "./src/electron/views/device-connect-view/components/android-setup/android-setup-spinner-step.tsx",
